requirements.txt
attrs==23.2.0
boto3==1.34.133
botocore==1.34.133
certifi==2024.6.2
h11==0.14.0
idna==3.7
jmespath==1.0.1
outcome==1.3.0.post0
pillow==10.3.0
psycopg2-binary==2.9.9
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.31
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
